# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projecte ETV",
        "description": "REST API. Projecte ETV. DAW Client i servidor.",
        "contact": {
            "name": "Isaac Palou.",
            "email": "isaacpalou@paucasesnovescifp.cat"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://www.etvtauladesfons.com/api"
        }
    ],
    "paths": {
...
        "/reserves/destroy/{id}": {
            "delete": {
                "tags": [
                    "Reserves"
                ],
                "summary": "Elimina una reserva",
                "description": "Elimina una reserva",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "result": {
                                            "type": "string",
                                            "example": "Reserva eliminada correctament"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "result": {
                                            "type": "string",
                                            "example": "No existeix aquesta reserva"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/usuaris": {
            "get": {
                "tags": [
                    "Usuaris"
                ],
                "summary": "Llista tots els usuaris",
                "operationId": "507a5d4f7f7dd00369273af78737b5dc",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "ID_USUARI": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "DNI": {
                                                        "type": "string",
                                                        "example": "12345678Z"
                                                    },
                                                    "NOM_COMPLET": {
                                                        "type": "string",
                                                        "example": "Pere Pujol"
                                                    },
                                                    "CORREU_ELECTRONIC": {
                                                        "type": "string",
                                                        "example": " patata@gmail.com"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuaris"
                ],
                "summary": "Crear un usuari",
                "operationId": "d2c3a53e08b4963ca026d128217ce859",
                "requestBody": {
                    "description": "Usuari a crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ID_USUARI": {
                                        "type": "integer",
                                        "example": "123"
                                    },
                                    "DNI": {
                                        "type": "string",
                                        "example": "12345678Z"
                                    },
                                    "NOM_COMPLET": {
                                        "type": "string",
                                        "example": "Pere Pujol"
                                    },
                                    "CORREU_ELECTRONIC": {
                                        "type": "string",
                                        "example": "perepujol@gmail.com"
                                    },
                                    "CONTRASENYA": {
                                        "type": "string",
                                        "example": "patata"
                                    },
                                    "TELEFON": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "ADMINISTRADOR": {
                                        "type": "boolean",
                                        "example": "false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "ID_USUARI": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "DNI": {
                                                        "type": "string",
                                                        "example": "12345678Z"
                                                    },
                                                    "NOM_COMPLET": {
                                                        "type": "string",
                                                        "example": "Pere Pujol"
                                                    },
                                                    "CORREU_ELECTRONIC": {
                                                        "type": "string",
                                                        "example": "patata@gmail.com"
                                                    },
                                                    "CONTRASENYA": {
                                                        "type": "string",
                                                        "example": "patata"
                                                    },
                                                    "TELEFON": {
                                                        "type": "string",
                                                        "example": "123456789"
                                                    },
                                                    "ADMINISTRADOR": {
                                                        "type": "boolean",
                                                        "example": "false"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autoritzat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/usuaris/{id}": {
            "get": {
                "tags": [
                    "Usuaris"
                ],
                "summary": "Mostra un usuari",
                "description": "Retorna un usuari",
                "operationId": "Obtenir usuari",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Usuari ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuari trobat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Usuari"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comentari no trobat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Usuari"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/usuaris/destroy/{id}": {
            "delete": {
                "tags": [
                    "Usuaris"
                ],
                "summary": "Elimina un usuari",
                "description": "Elimina un usuari",
                "operationId": "f5e0e90c10a1812914deffc79efa2d4e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'usuari",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuari eliminat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Usuari"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuari no trobat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Usuari"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de usuari",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Usuari"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Usuari": {
                "properties": {
                    "ID_USUARI": {
                        "type": "integer",
                        "example": 1
                    },
                    "DNI": {
                        "type": "string",
                        "example": "12345678Z"
                    },
                    "NOM_COMPLET": {
                        "type": "string",
                        "example": "Pere Pujol"
                    },
                    "CORREU_ELECTRONIC": {
                        "type": "string",
                        "example": "patata@gmail.com"
                    },
                    "CONTRASENYA": {
                        "type": "string",
                        "example": "12345678Z"
                    },
                    "TELEFON": {
                        "type": "string",
                        "example": "12345678Z"
                    },
                    "ADMINISTRADOR": {
                        "type": "boolean",
                        "example": "true"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}